requirements/requirements-dev.txt
pylint
pycodestyle
tox
pytest
pytest-cov
pytest-env
pytest-django
